[refactor] admin_list 상품 삭제 버튼 클릭 시 알림 창을 띄워 다시 한번 물어보는 기능 추가

+++ src/main/webapp/WEB-INF/views/admin_list.jsp
@@ -40,7 +40,7 @@
         sb.append("<td>");
         sb.append("<input type=\"button\" value=\"상품수정\" class=\"btn_write btn_txt01\" style=\"cursor: pointer;\" onclick=\"location.href='./modify?productId=" + productId + "'\" />");
         sb.append("&nbsp;&nbsp;");
-        sb.append("<input type=\"button\" value=\"상품삭제\" class=\"btn_write btn_txt01\" style=\"cursor: pointer;\" onclick=\"location.href='./delete?productId=" + productId + "'\" />");
+        sb.append("<input type=\"button\" value=\"상품삭제\" class=\"btn_write btn_txt01\" style=\"cursor: pointer;\" onclick=\"if(confirm('정말 삭제하시겠습니까?')) { location.href='./delete?productId=" + productId + "'; }\" />");
         sb.append("</td>");
         sb.append("</tr>");
     }
